--- src/pages/Article/components/Comment/index.ts
@@ -1,3 +1,0 @@
-import Comment from './Comment';
-
-export default Comment;
