--- src/pages/Article/components/Comment/style.scss
@@ -1,75 +1,0 @@
-.comment,
-.reply {
-	display: flex;
-	align-items: flex-start;
-	word-wrap: break-word;
-
-	.avatar {
-		margin-right: 10px;
-	}
-
-	.content {
-		flex: 1;
-		padding: 0;
-
-		.info {
-			display: flex;
-			align-items: center;
-			margin-bottom: 5px;
-
-			.username {
-				margin-right: 10px;
-			}
-
-			.date {
-				opacity: 0.6;
-			}
-		}
-
-		.text {
-			margin-bottom: 10px;
-		}
-
-		.actions {
-			display: flex;
-			align-items: center;
-
-			.assessment {
-				display: flex;
-				align-items: center;
-				margin-right: 10px;
-
-				.likes-number {
-					margin-right: 10px;
-				}
-			}
-		}
-	}
-
-	@media (max-width: 400px) {
-		.info {
-			flex-direction: column;
-			align-items: flex-start !important;
-
-			.username {
-				margin-right: 0;
-			}
-		}
-	}
-}
-
-.comment {
-	margin-bottom: 20px;
-
-	&:last-child {
-		margin-bottom: 0;
-	}
-}
-
-.sub-comment {
-	margin-bottom: 10px;
-
-	&:last-child {
-		margin-bottom: 0;
-	}
-}
