+++ src/pages/Article/components/CommentReplies/index.ts
@@ -1,0 +1,3 @@
+import CommentReplies from './CommentReplies';
+
+export default CommentReplies;
