+++ src/pages/UsersChat/style.scss
@@ -2,4 +2,5 @@
 	display: flex;
 	flex-direction: column;
 	flex: 1;
+	position: relative;
 }
