+++ src/utils/callApi.ts
@@ -1,10 +1,28 @@
-import axios from 'axios';
+import axios, {AxiosRequestConfig} from 'axios';
 
 import {handleNetworkError} from './errorHandlers';
 
-const callApi = async (endpoint: string, options: object = {}): Promise<any> => {
+const getToken = (): string => localStorage.getItem('token') || '';
+
+const request = axios.create({
+	baseURL: String(process.env.REACT_APP_API),
+	headers: {
+		Authorization: `Bearer ${getToken()}`,
+	},
+});
+
+request.interceptors.request.use(
+	config => {
+		config.headers.Authorization = `Bearer ${getToken()}`;
+
+		return config;
+	},
+	error => handleNetworkError(error),
+);
+
+const callApi = async (endpoint: string, options?: AxiosRequestConfig): Promise<any> => {
 	try {
-		const {data} = await axios(`${process.env.REACT_APP_API}/${endpoint}`, options);
+		const {data} = await request(endpoint, options);
 
 		return data;
 	} catch (error) {
@@ -12,9 +30,9 @@ const callApi = async (endpoint: string, options: object = {}): Promise<any> => 
 	}
 };
 
-callApi.get = async (endpoint: string, options: object = {}): Promise<any> => {
+callApi.get = async (endpoint: string, options?: AxiosRequestConfig): Promise<any> => {
 	try {
-		const {data} = await axios.get(`${process.env.REACT_APP_API}/${endpoint}`, options);
+		const {data} = await request.get(endpoint, options);
 
 		return data;
 	} catch (error) {
@@ -25,10 +43,10 @@ callApi.get = async (endpoint: string, options: object = {}): Promise<any> => {
 callApi.post = async (
 	endpoint: string,
 	payload: object = {},
-	options: object = {},
+	options?: AxiosRequestConfig,
 ): Promise<any> => {
 	try {
-		const {data} = await axios.post(`${process.env.REACT_APP_API}/${endpoint}`, payload, options);
+		const {data} = await request.post(endpoint, payload, options);
 
 		return data;
 	} catch (error) {
@@ -39,10 +57,10 @@ callApi.post = async (
 callApi.put = async (
 	endpoint: string,
 	payload: object = {},
-	options: object = {},
+	options?: AxiosRequestConfig,
 ): Promise<any> => {
 	try {
-		const {data} = await axios.put(`${process.env.REACT_APP_API}/${endpoint}`, payload, options);
+		const {data} = await request.put(endpoint, payload, options);
 
 		return data;
 	} catch (error) {
@@ -50,9 +68,9 @@ callApi.put = async (
 	}
 };
 
-callApi.delete = async (endpoint: string, options: object = {}): Promise<any> => {
+callApi.delete = async (endpoint: string, options?: AxiosRequestConfig): Promise<any> => {
 	try {
-		const {data} = await axios.delete(`${process.env.REACT_APP_API}/${endpoint}`, options);
+		const {data} = await request.delete(endpoint, options);
 
 		return data;
 	} catch (error) {
@@ -61,34 +79,3 @@ callApi.delete = async (endpoint: string, options: object = {}): Promise<any> =>
 };
 
 export default callApi;
-
-// --------------------------------------------------------------------------------------
-// const identity = <T>(arg: Array<T>): Array<T> => {
-// 	return arg;
-// };
-
-// const result = identity<number>([1, 2, 3]);
-// console.log(result.length);
-
-// function getBy<T, P extends keyof T>(model: T[], prop: P, value: T[P]): T | null {
-// 	return model.filter(item => item[prop] === value)[0] || null;
-// }
-// --------------------------------------------------------------------------------------
-// const result = getBy(students, 'age', '17');
-// // Error: Argument of type '"17"' is not assignable to parameter of type 'number'.
-
-// const anotherResult = getBy(students, 'hasScar', 'true');
-// // Error: Argument of type '"true"' is not assignable to parameter of type 'boolean'.
-
-// const yetAnotherResult = getBy(students, 'name', 'Harry');
-// // А тут уже всё правильно
-// --------------------------------------------------------------------------------------
-// callApi.get = async <ReturningData>(
-// 	endpoint: string,
-// 	options: object = {},
-// ): Promise<ReturningData> => {
-// 	const {data} = await axios.get(`${process.env.REACT_APP_API}/${endpoint}`, options);
-
-// 	return data;
-// };
-// --------------------------------------------------------------------------------------
